requirements.txt
fdk
pyjwt[crypto]
pyOpenSSL 
oci
requests
